+++ c/tskit/trees.c
@@ -492,15 +492,14 @@ tsk_treeseq_copy_tables(
 }
 
 int TSK_WARN_UNUSED
-tsk_treeseq_load(
-    tsk_treeseq_t *self, const char *filename, tsk_flags_t TSK_UNUSED(options))
+tsk_treeseq_load(tsk_treeseq_t *self, const char *filename, tsk_flags_t options)
 {
     int ret = 0;
     tsk_table_collection_t tables;
 
     /* Need to make sure that we're zero'd out in case of error */
     tsk_memset(self, 0, sizeof(*self));
-    ret = tsk_table_collection_load(&tables, filename, 0);
+    ret = tsk_table_collection_load(&tables, filename, options);
     if (ret != 0) {
         goto out;
     }
@@ -518,14 +517,14 @@ out:
 }
 
 int TSK_WARN_UNUSED
-tsk_treeseq_loadf(tsk_treeseq_t *self, FILE *file, tsk_flags_t TSK_UNUSED(options))
+tsk_treeseq_loadf(tsk_treeseq_t *self, FILE *file, tsk_flags_t options)
 {
     int ret = 0;
     tsk_table_collection_t tables;
 
     /* Need to make sure that we're zero'd out in case of error */
     tsk_memset(self, 0, sizeof(*self));
-    ret = tsk_table_collection_loadf(&tables, file, 0);
+    ret = tsk_table_collection_loadf(&tables, file, options);
     if (ret != 0) {
         goto out;
     }